Removed a bunch of dependencies, depth buffer stuff
Clean-up on a whole bunch of dependencies that are no longer in use.
Started adding support for depth cameras for whenever I figure out how to hook up the the intel depthsense cam.
Commented out VTTexture for the time being since it was the only thing using some of the dependencies I cleaned up.